ホームに特有のコードをHomeFeedFragmentにわける
